requirements.txt
# Core
PyYAML>=6.0
pydantic>=2.7
pydantic-settings>=2.2
requests>=2.32
httpx>=0.27
tenacity>=8.3

# Data processing
pandas>=2.2
numpy>=1.26

# Geospatial and environmental data
geopandas>=1.0
rasterio>=1.3
shapely>=2.0
xarray>=2024.2

# Scheduling and workflows
APScheduler>=3.10

# API / service layer
fastapi>=0.111
uvicorn[standard]>=0.30

# Observability
structlog>=24.1
python-json-logger>=2.0

# Testing and quality
pytest>=8.2
pytest-cov>=5.0
ruff>=0.5

